requirements.txt
Django==6.0
python-decouple==3.8
